# docs/swagger.yaml
definitions:
...
  json.AvaliacaoEstatisticasResponse:
    properties:
      media_nota:
        type: number
      nota_1:
        type: integer
      nota_2:
        type: integer
      nota_3:
        type: integer
      nota_4:
        type: integer
      nota_5:
        type: integer
      total_avaliacoes:
        type: integer
    type: object
  json.AvaliacaoRequest:
    properties:
      comentario:
        maxLength: 500
        type: string
      id_loja:
        type: integer
      id_usuario:
        type: integer
      nota:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - id_loja
    - id_usuario
    - nota
    type: object
  json.AvaliacaoResponse:
    properties:
      comentario:
        type: string
      data_atualizacao:
        type: string
      data_avaliacao:
        type: string
      id:
        type: integer
      id_loja:
        type: integer
      id_usuario:
        type: integer
      loja:
        $ref: '#/definitions/json.LojaResponse'
      nota:
        type: integer
      usuario:
        allOf:
        - $ref: '#/definitions/json.UserResponse'
        description: Dados relacionados
    type: object
  json.AvaliacoesResponse:
    properties:
      avaliacoes:
        items:
          $ref: '#/definitions/json.AvaliacaoResponse'
        type: array
      media_nota:
        type: number
      total:
        type: integer
    type: object
...
  json.LojaResponse:
    properties:
      categoria:
        type: string
      cnpj:
        type: string
      distancia:
        description: Distância em km
        type: number
      id:
        type: integer
      id_categoria:
        type: integer
      imagem:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      nome:
        type: string
    type: object
...
  json.UserResponse:
    properties:
      ativo:
        type: boolean
      cpf:
        type: string
      data_cadastro:
        type: string
      data_nascimento:
        type: string
      email:
        type: string
      endereco:
        type: string
      id:
        type: integer
      id_loja:
        type: integer
      id_plano:
        type: integer
      imagem:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      mensagem:
        type: string
      nome:
        type: string
      telefone:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /avaliacoes:
    get:
      consumes:
      - application/json
      description: Retorna todas as avaliações ativas do sistema
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.AvaliacoesResponse'
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista todas as avaliações
      tags:
      - Avaliações
    post:
      consumes:
      - application/json
      description: Cria uma nova avaliação de loja no sistema
      parameters:
      - description: Dados da avaliação
        in: body
        name: avaliacao
        required: true
        schema:
          $ref: '#/definitions/json.AvaliacaoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/json.AvaliacaoResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Cria uma nova avaliação
      tags:
      - Avaliações
  /avaliacoes/{id}:
    delete:
      consumes:
      - application/json
      description: Realiza soft delete de uma avaliação, marcando-a como excluída
      parameters:
      - description: ID da avaliação
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Avaliação removida com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Avaliação não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Remove avaliação (soft delete)
      tags:
      - Avaliações
    get:
      consumes:
      - application/json
      description: Retorna uma avaliação específica pelo ID
      parameters:
      - description: ID da avaliação
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.AvaliacaoResponse'
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Avaliação não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Busca avaliação por ID
      tags:
      - Avaliações
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma avaliação existente
      parameters:
      - description: ID da avaliação
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados da avaliação
        in: body
        name: avaliacao
        required: true
        schema:
          $ref: '#/definitions/json.AvaliacaoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.AvaliacaoResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Avaliação não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza avaliação
      tags:
      - Avaliações
  /avaliacoes/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restaura uma avaliação que foi soft deleted
      parameters:
      - description: ID da avaliação
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Avaliação restaurada com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Avaliação não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Restaura avaliação
      tags:
      - Avaliações
...
  /lojas/{id}/avaliacoes:
    get:
      consumes:
      - application/json
      description: Retorna todas as avaliações de uma loja específica
      parameters:
      - description: ID da loja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.AvaliacoesResponse'
        "400":
          description: ID de loja inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista avaliações de uma loja
      tags:
      - Avaliações
  /lojas/{id}/avaliacoes/estatisticas:
    get:
      consumes:
      - application/json
      description: Retorna estatísticas detalhadas das avaliações de uma loja
      parameters:
      - description: ID da loja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.AvaliacaoEstatisticasResponse'
        "400":
          description: ID de loja inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Estatísticas de avaliações de uma loja
      tags:
      - Avaliações
...
  /users/{id}/avaliacoes:
    get:
      consumes:
      - application/json
      description: Retorna todas as avaliações feitas por um usuário específico
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.AvaliacoesResponse'
        "400":
          description: ID de usuário inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista avaliações de um usuário
      tags:
      - Avaliações
...
swagger: "2.0"